+++ internal/provider/provider_data.go
@@ -3,6 +3,7 @@ package provider
 type HostProviderData struct {
 	HomeDir               string
 	RuntimeDir            string
+	TargetUser            string
 	PackageManager        PackageManager
 	BrewManager           BrewPackageManager
 	ScheduleManager       ScheduleManager